first iteration of possible change, adding in z-index for sticky nav styling, use of project component for portfolio page, and centered carousel on d-flex
