+++ src/components/project.js
@@ -1,5 +1,8 @@
 // packages
-import React from "react";
+import React, { useState } from "react";
+
+// import bootstrap components
+import { Carousel } from "react-bootstrap";
 
 // // assets
 // import inorout from "../assets/inorout.png";
@@ -13,11 +16,48 @@ import React from "react";
 // };
 
 function Project({ data }) {
+  // hook for carousel
+  const [index, setIndex] = useState(0);
+  const handleSelect = (selectedIndex, e) => {
+    setIndex(selectedIndex);
+  };
+
   return (
-    <div>
-      <h1>Project 1</h1>
-      
-    </div>
+    <>
+      <div className="d-flex justify-content-center">
+        <Carousel
+          activeIndex={index}
+          onSelect={handleSelect}
+          className="carousel-margin-top carousel-width"
+        >
+          <Carousel.Item>
+            <img
+              className="d-block w-100"
+              src="https://via.placeholder.com/800x400"
+              alt="first project slide"
+            />
+            <Carousel.Caption>
+              <h3>Project 1 slide 1</h3>
+            </Carousel.Caption>
+          </Carousel.Item>
+          <Carousel.Item>
+            <img
+              className="d-block w-100"
+              src="https://via.placeholder.com/800x400"
+              alt="second project slide"
+            />
+            <Carousel.Caption>
+              <h3>Project 1 slide 2</h3>
+            </Carousel.Caption>
+          </Carousel.Item>
+        </Carousel>
+      </div>
+      <div >
+        <p>Description of app</p>
+        <p>Link to github</p>
+        <p>Link to application</p>
+      </div>
+    </>
   );
 }
