+++ src/style.css
@@ -93,6 +93,7 @@
   position: fixed;
   top: 0;
   width: 100%;
+  z-index: 100;
 }
 
 /* footer */
@@ -119,9 +120,12 @@
 }
 
 /* carousel styling */
-.carousel-height {
-  width: 100%;
-  height: 500px;
+.carousel-width {
+  width: 90%
+}
+
+.img-height {
+  height: 90%;
 }
 
 .carousel-margin-top {
